backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.25
mysql-connector-python==8.3.0

pydantic==2.6.4
pydantic-settings==2.2.1

python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2

python-dotenv==1.0.1
python-multipart==0.0.9
